feat: optimize the gantt zoom
