+++ packages/vtable-gantt/src/zoom-scale/ZoomScaleManager.ts
@@ -2,6 +2,7 @@ import type { Gantt } from '../Gantt';
 import type { IZoomScale } from '../ts-types/zoom-scale';
 import type { ITimelineScale } from '../ts-types/gantt-engine';
 import { DataZoomIntegration, type DataZoomConfig } from './DataZoomIntegration';
+import { GANTT_EVENT_TYPE } from '../ts-types/EVENT_TYPE';
 
 /**
  * ZoomScale 管理器
@@ -540,6 +541,137 @@ export class ZoomScaleManager {
   updateDataZoomResponsive(): void {
     if (this.dataZoomIntegration) {
       this.dataZoomIntegration.updateResponsive();
+    }
+  }
+
+  /**
+   * 重新计算时间相关的尺寸参数
+   * 用于根据当前 timePerPixel 重新计算 timelineColWidth
+   * （从 Gantt.ts 移动到这里）
+   */
+  recalculateTimeScale(): void {
+    // 获取当前的主时间刻度
+    const primaryScale = this.gantt.parsedOptions.reverseSortedTimelineScales[0];
+    if (!primaryScale) {
+      return;
+    }
+
+    // 根据当前 scale 的 unit 和 step 计算每个单元格应该占用的毫秒数
+    let msPerStep: number;
+    switch (primaryScale.unit as string) {
+      case 'second':
+        msPerStep = 1000 * primaryScale.step;
+        break;
+      case 'minute':
+        msPerStep = 60 * 1000 * primaryScale.step;
+        break;
+      case 'hour':
+        msPerStep = 60 * 60 * 1000 * primaryScale.step;
+        break;
+      case 'day':
+        msPerStep = 24 * 60 * 60 * 1000 * primaryScale.step;
+        break;
+      case 'week':
+        msPerStep = 7 * 24 * 60 * 60 * 1000 * primaryScale.step;
+        break;
+      case 'month':
+        msPerStep = 30 * 24 * 60 * 60 * 1000 * primaryScale.step;
+        break;
+      case 'quarter':
+        msPerStep = 90 * 24 * 60 * 60 * 1000 * primaryScale.step;
+        break;
+      case 'year':
+        msPerStep = 365 * 24 * 60 * 60 * 1000 * primaryScale.step;
+        break;
+      default:
+        msPerStep = 24 * 60 * 60 * 1000 * primaryScale.step;
+    }
+    const newTimelineColWidth = msPerStep / this.gantt.getCurrentTimePerPixel();
+
+    this.gantt.parsedOptions.timelineColWidth = newTimelineColWidth;
+
+    // 重新生成时间线日期映射
+    this.gantt._generateTimeLineDateMap();
+
+    // 更新尺寸和重新渲染
+    if (this.gantt.scenegraph) {
+      this.gantt._updateSize();
+      this.gantt.scenegraph.refreshAll();
+    }
+  }
+
+  /**
+   * 缩放方法，用于滚轮和双指缩放
+   * （从 Gantt.ts 移动到这里）
+   * @param factor 缩放因子，大于1表示放大
+   * @param keepCenter 是否保持视图中心不变
+   * @param centerX 缩放中心点X坐标
+   */
+  zoomByFactor(factor: number, keepCenter: boolean = true, centerX?: number): void {
+    const minTimePerPixel = this.gantt.parsedOptions.zoom?.minTimePerPixel ?? 200000;
+    const maxTimePerPixel = this.gantt.parsedOptions.zoom?.maxTimePerPixel ?? 3000000;
+
+    const oldTimePerPixel = this.gantt.getCurrentTimePerPixel();
+    const oldWidth = this.gantt.parsedOptions.timelineColWidth;
+
+    // 在级别切换前先计算中心时间位置
+    let centerTimePosition: number | undefined;
+    if (keepCenter) {
+      if (centerX === undefined) {
+        centerX = this.gantt.scenegraph.width / 2;
+      }
+      // 计算中心点对应的绝对时间位置
+      const scrollOffsetMs = (this.gantt.stateManager.scroll.horizontalBarPos + centerX) * oldTimePerPixel;
+      centerTimePosition = this.gantt.parsedOptions._minDateTime + scrollOffsetMs;
+    }
+
+    const currentTimePerPixel = this.gantt.getCurrentTimePerPixel();
+    let adjustedFactor = factor;
+
+    const baseTimePerPixel = 1440000;
+    const zoomRatio = Math.log(currentTimePerPixel / baseTimePerPixel) / Math.log(2);
+
+    if (currentTimePerPixel < baseTimePerPixel) {
+      const enhancement = Math.pow(1.2, -zoomRatio);
+      adjustedFactor = Math.pow(factor, enhancement);
+    } else {
+      const dampening = Math.pow(0.9, zoomRatio);
+      adjustedFactor = Math.pow(factor, dampening);
+    }
+
+    const newTimePerPixel = this.gantt.getCurrentTimePerPixel() / adjustedFactor;
+    const clampedTimePerPixel = Math.max(minTimePerPixel, Math.min(maxTimePerPixel, newTimePerPixel));
+    this.gantt.setTimePerPixel(clampedTimePerPixel);
+
+    const targetLevel = this.findOptimalLevel(clampedTimePerPixel);
+    const currentLevel = this.getCurrentLevel();
+
+    if (targetLevel !== currentLevel) {
+      this.switchToLevel(targetLevel);
+    } else {
+      this.recalculateTimeScale();
+    }
+
+    // 在级别切换和重新计算后再调整视图中心
+    if (keepCenter && centerTimePosition !== undefined && centerX !== undefined) {
+      const actualTimePerPixel = this.gantt.getCurrentTimePerPixel();
+      // 计算中心时间相对于新minDate的偏移量
+      const newMinDateTime = this.gantt.parsedOptions._minDateTime;
+      const timeOffsetFromNewMin = centerTimePosition - newMinDateTime;
+      const newScrollLeft = timeOffsetFromNewMin / actualTimePerPixel - centerX;
+      this.gantt.stateManager.setScrollLeft(newScrollLeft);
+    }
+
+    // 触发缩放事件
+    if (this.gantt.hasListeners(GANTT_EVENT_TYPE.ZOOM)) {
+      const actualTimePerPixel = this.gantt.getCurrentTimePerPixel();
+      this.gantt.fireListeners(GANTT_EVENT_TYPE.ZOOM, {
+        oldWidth,
+        newWidth: this.gantt.parsedOptions.timelineColWidth,
+        scale: oldTimePerPixel / actualTimePerPixel,
+        oldTimePerPixel,
+        newTimePerPixel: actualTimePerPixel
+      });
     }
   }
 }
